Implement toggling of system device display
't' now toggles the display of system devices, which are non-removable
devices like hard disks.

Since 't' generically stands for 'toggle', the key bindings may be ill
chosen and could be subject to change in the future. 's' for 'system'
should be reserved for 'open shell in mount path', since that is
probably the more intuitive binding.